Adding get_connection to separate file.
